// apps/enterprise-service-management/src/generated/models/lifecycle-definition-model.ts
import type { Workspace } from './workspace-model';

export const LifecycleDefinitionStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type LifecycleDefinitionStatusKey = keyof typeof LifecycleDefinitionStatusKeyToLabel;

export interface LifecycleDefinition {
  /**
   * @displayName Lifecycle Definition ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  name1: string;
  /**
   * @displayName Catalog Item Code
   */
  catalogItemCode?: string;
  /**
   * @displayName Code
   * @validationRule Required for create/update operations
   */
  code: string;
  /**
   * @displayName Default
   * @validationRule Required for create/update operations
   */
  default1: boolean;
  /**
   * @displayName Request Type Code
   * @validationRule Required for create/update operations
   */
  requestTypeCode: string;
  /**
   * @displayName Service Code
   */
  serviceCode?: string;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: LifecycleDefinitionStatusKey;
  /**
   * @displayName Version
   * @validationRule Required for create/update operations
   */
  version: number;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _LifecycleDefinition = 'LifecycleDefinition' as const;