// src/generated/models/status-definition-model.ts
import type { LifecycleDefinition } from './lifecycle-definition-model';
import type { LifecycleStage } from './lifecycle-stage-model';
import type { Workspace } from './workspace-model';

export const StatusDefinitionStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type StatusDefinitionStatusKey = keyof typeof StatusDefinitionStatusKeyToLabel;

export interface StatusDefinition {
  /**
   * @displayName Status Definition ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Status Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  statusName: string;
  /**
   * @displayName Is Initial
   * @validationRule Required for create/update operations
   */
  isInitial: boolean;
  /**
   * @displayName Is Terminal
   * @validationRule Required for create/update operations
   */
  isTerminal: boolean;
  /**
   * @displayName Lifecycle Definition ID
   * @validationRule Required for create/update operations
   */
  lifecycleDefinitionID: Pick<LifecycleDefinition, 'id' | 'name1'>;
  /**
   * @displayName Lifecycle Stage ID
   * @validationRule Required for create/update operations
   */
  lifecycleStageID: Pick<LifecycleStage, 'id' | 'stageName'>;
  /**
   * @displayName Sort Order
   * @validationRule Required for create/update operations
   */
  sortOrder: number;
  /**
   * @displayName Status Code
   * @validationRule Required for create/update operations
   */
  statusCode: string;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: StatusDefinitionStatusKey;
  /**
   * @displayName Workspace ID
   * @validationRule Required for create/update operations
   */
  workspaceID: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _StatusDefinition = 'StatusDefinition' as const;